Make tachyon-client only depend on tachyon-common by moving DataServerMessage to common and define MASTER_INFO_COL in Constants
