+++ client/src/main/java/tachyon/client/table/RawColumn.java
@@ -17,10 +17,10 @@ package tachyon.client.table;
 
 import java.io.IOException;
 
+import tachyon.Constants;
 import tachyon.TachyonURI;
 import tachyon.client.TachyonFile;
 import tachyon.client.TachyonFS;
-import tachyon.master.MasterInfo;
 import tachyon.util.CommonUtils;
 
 /**
@@ -45,8 +45,8 @@ public class RawColumn {
   // TODO creating file here should be based on id.
   public boolean createPartition(int pId) throws IOException {
     TachyonURI tUri =
-        new TachyonURI(
-            CommonUtils.concatPath(mRawTable.getPath(), MasterInfo.COL + mColumnIndex, pId));
+        new TachyonURI(CommonUtils.concatPath(mRawTable.getPath(), Constants.MASTER_INFO_COL
+            + mColumnIndex, pId));
     return mTachyonFS.createFile(tUri) > 0;
   }
 
@@ -58,15 +58,16 @@ public class RawColumn {
   // TODO creating file here should be based on id.
   public TachyonFile getPartition(int pId, boolean cachedMetadata) throws IOException {
     TachyonURI tUri =
-        new TachyonURI(
-            CommonUtils.concatPath(mRawTable.getPath(), MasterInfo.COL + mColumnIndex, pId));
+        new TachyonURI(CommonUtils.concatPath(mRawTable.getPath(), Constants.MASTER_INFO_COL
+            + mColumnIndex, pId));
     return mTachyonFS.getFile(tUri, cachedMetadata);
   }
 
   // TODO creating file here should be based on id.
   public int partitions() throws IOException {
     TachyonURI tUri =
-        new TachyonURI(CommonUtils.concatPath(mRawTable.getPath(), MasterInfo.COL + mColumnIndex));
+        new TachyonURI(CommonUtils.concatPath(mRawTable.getPath(), Constants.MASTER_INFO_COL
+            + mColumnIndex));
     return mTachyonFS.listStatus(tUri).size();
   }
 }
