+++ common/src/main/java/tachyon/Constants.java
@@ -206,4 +206,6 @@ public class Constants {
 
   public static final String S3_ACCESS_KEY = "fs.s3n.awsAccessKeyId";
   public static final String S3_SECRET_KEY = "fs.s3n.awsSecretAccessKey";
+
+  public static final String MASTER_INFO_COL = "COL_";
 }
