+++ core/pom.xml
@@ -20,7 +20,7 @@
   <dependencies>
     <dependency>
       <groupId>org.tachyonproject</groupId>
-      <artifactId>tachyon-common</artifactId>
+      <artifactId>tachyon-client</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
